Add composer, commonmark parser
There is now basic support for md for instructions and for intro
sections. However, there will be some discrepancies between the look of
instructions sections in existing recipes. Also, the CSS should be
updated to handle the HTML that is rendered now. E.g. recognise headers
and lists and style them appropriately.