Disable SimpleMDE autosave
This fixes the bug where when opening a recipe to edit, the
"instructions" text area was pre-filled with text from other recipes.
The bug was actually that `uniqueId` was not unique for each
textarea-recipe pair (all had uniqueId set to 'instructions').

The solution was to set a proper unique id, however, I chose to disable
the functionality altogether as the feature would save local edits
forever until submitted thus overriding remote changes from other
clients.